requirements.txt
google-genai
opencv-python
matplotlib
scipy
numpy
